@keegancsmith 2021-07-29
Surprisingly I didn't notice a difference on CI. I am guessing the ephemeralDB setup failed. I will debug further. But this worked really well for my devenv. Interested to see impact for other if they wanna try out this branch before and after. `go test -count=1 ./internal/database`