@keegancsmith 2021-07-29
Pushed a commit to support this on CI. `internal/database` goes from 116s to 39s :)