@keegancsmith 2021-08-05
The reason it is fast for you is you turned off fsync. Note: this is a manual step you did.  FYI we don't mention turning it off anywhere in our dev docs. Here are the timings on two of my machines. The "old" includes my system postgres which hasn't been tuned (so fsync on).

On my macbook pro we go from ~27s to ~22s:

![image](https://user-images.githubusercontent.com/187831/128371756-8a1d8daa-44a2-472f-89f4-980940ec72ff.png)

On my linux desktop (with ext4 + Samsung SSD 980 PRO 2TB) we go from ~1m10s to ~9s.

![1628175382](https://user-images.githubusercontent.com/187831/128372013-2bdeee62-3b92-4fdc-9541-01775531f83d.png)

I think the complexity is not that bad to be honest. A lot of it is due to CI and it running as root. This can be removed by improving CI, but I tried to make it not bloat the code too much.

Philisophically I really like this approach. If we extend it to our whole dev environment we remove the need for a user to setup postgres. IE that bit of onboarding goes away + we can control the tuning to ensure performance. From that perspective what do yo uthink? @mrnugget