@mrnugget 2021-08-06
> Philisophically I really like this approach. If we extend it to our whole dev environment we remove the need for a user to setup postgres. IE that bit of onboarding goes away + we can control the tuning to ensure performance. From that perspective what do yo uthink? @mrnugget

Yeah, I'd like that. But what I'm skeptical about is that this is "hidden" in a `TestMain`.

To be clear: I'm not explicitly blocking this PR.

But what I'd like more:

1. Docs that show how to tune local DB performance
2. A check for these settings when starting dev-env/tests that lets the user know "hey, you're running unoptimized. read X to tune". (If we decide to all adopt `sg` we could add a check to `sg.config.yaml` that does that. Maybe something like "soft checks" that only print warnings)
3. One of the focuses in Q3 is improving the dev onboarding experience. One of the ideas we've been brainstorming was to have something like `sg setup` which you can run on a fresh machine and that guides you through the installation of dev dependencies ("Postgres. Choose: (a) Docker (b) Postgres.app (c) homebrew (d) apt") and sets up your sg config accordingly. That, I think, would be the ideal place to nail down these recommendations.