@keegancsmith 2021-08-06
> Yeah, I'd like that. But what I'm skeptical about is that this is "hidden" in a `TestMain`.

This is intentional to demonstrate the idea. If feedback is positive I'm happy to apply this across our test codebase as well as devenv.

> But what I'd like more:

Can you expand on why you prefer your approach more? From my perspective why have a multi step approach when the dev can just do nothing at all. My proposal is if they use something like ASDF they don't need to do _anything_ at all, it just works.

Fresh macbook/linux/etc. Install asdf (and probably docker). You can now successfully just run `go test` and `sg start` without needing to anything more. We guarentee the correct versions of postgres/redis/etc. No need to set env vars, run initdb commands, etc. All state lives (including sourcegraph service state) live under a common directory.