@mrnugget 2021-08-06
> Can you expand on why you prefer your approach more? From my perspective why have a multi step approach

Sorry, I was unclear: it wasn't meant as a multi-step thing, but rather "these are all options I think are good".

Why do I like those more? Because they're explicit and I can already see myself having to debug some tests and forgetting that somewhere else we setup an internal database for these tests that I can't reach from outside, etc. 

> Fresh macbook/linux/etc. Install asdf (and probably docker). You can now successfully just run `go test` and `sg start` without needing to anything more. We guarentee the correct versions of postgres/redis/etc. No need to set env vars, run initdb commands, etc. All state lives (including sourcegraph service state) live under a common directory.

That to me sounds orthogonal to (3) I mentioned.