@keegancsmith 2021-10-06
I've since started using nix and made it so our `shell.nix` sets up a postgres optimized for dev. I still think there is merit in making our DB tests not depend on a functioning postgres (but just the postgres binaries), but gonna close this out since there wasn't enough buy in :)